requirements.txt
flask
gunicorn
pillow
requests
numpy
torch
torchvision
rembg
onnxruntime
